fix subtle comment detection bug
For stuff like \email{whatever}, the part "{whatever}" is in verbatim
font, but we were just checking this on the first characters of the
delimiters.  Checking it on the final characters seems to work.  A bit
inelegant since it means that we ignore such macros, but don't see any
way around doing so without resorting to "non-local" parsing.