eliminate dependence on `transpose-sexps-function'
This variable was introduced only in Emacs 30.1, and is not needed for
the package to work, so we have removed reference to it, replacing its
role with a new defun `tex-parens-transpose-sexps'.  As a consequence,
`tex-parens-transpose-sexps' now needs to be bound in the keymap
underlying `tex-parens-mode' and also in any user config (see the README).